Fixed issues with yandex merchant

+++ Dao/src/main/java/me/exrates/dao/WalletDao.java
@@ -1,9 +1,8 @@
 package me.exrates.dao;
 
-import java.math.BigDecimal;
-import java.util.List;
-
 import me.exrates.model.Wallet;
+
+import java.util.List;
 
 public interface WalletDao {
 
@@ -11,9 +10,9 @@ public interface WalletDao {
 
 	double getWalletRBalance(int walletId);
 
-	boolean setWalletABalance(int walletId, BigDecimal newBalance);
+	boolean setWalletABalance(int walletId, double newBalance);
 
-	boolean setWalletRBalance(int walletId, BigDecimal newBalance);
+	boolean setWalletRBalance(int walletId, double newBalance);
 
 	int getWalletId(int userId, int currencyId);
